[svn-r25496] Description:     Merge 64-bit ID changes from branch to trunk.  (Plus a few minor cleanups that aren't on the branch)
Tested on:
    Mac OSX/64 10.9.4 (amazon) w/C++ & FORTRAN
    (h5committested on branch already for a week)

+++ src/H5Iprivate.h
@@ -66,7 +66,7 @@ typedef struct H5I_class_t {
 /* Library-private Function Prototypes */
 /***************************************/
 H5_DLL herr_t H5I_register_type(const H5I_class_t *cls);
-H5_DLL int H5I_nmembers(H5I_type_t type);
+H5_DLL int64_t H5I_nmembers(H5I_type_t type);
 H5_DLL herr_t H5I_clear_type(H5I_type_t type, hbool_t force, hbool_t app_ref);
 H5_DLL hid_t H5I_register(H5I_type_t type, const void *object, hbool_t app_ref);
 H5_DLL void *H5I_subst(hid_t id, const void *new_object);
